+++ Source/Extensions.StringSegment.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics;
-using System.Text;
+﻿using System.Text;
 
 namespace Open.Text;
