+++ Source/Extensions._.cs
@@ -8,8 +8,10 @@ public static partial class TextExtensions
 {
 	private const uint BYTE_RED = 1024;
 
-	private static IEnumerable<string> ByteLabels {
-		get {
+	private static IEnumerable<string> ByteLabels
+	{
+		get
+		{
 			yield return "KB";
 			yield return "MB";
 			yield return "GB";
@@ -137,34 +139,6 @@ public static partial class TextExtensions
 	public static bool IsAlphaNumeric(this string source, bool trim = false)
 		=> !string.IsNullOrWhiteSpace(source)
 		&& (trim ? RegexPatterns.ValidAlphaNumericOnlyUntrimmedPattern : RegexPatterns.ValidAlphaNumericOnlyPattern).IsMatch(source);
-
-	/// <summary>
-	/// Returns the available matches as StringSegments.
-	/// </summary>
-	/// <param name="pattern">The pattern to search with.</param>
-	/// <param name="input">The string to search.</param>
-	/// <returns>An enumerable containing the found segments.</returns>
-	/// <exception cref="ArgumentNullException">If the pattern or input is null.</exception>
-	public static IEnumerable<StringSegment> AsSegments(this Regex pattern, string input)
-	{
-		return pattern is null
-			? throw new ArgumentNullException(nameof(pattern))
-			: input is null
-			? throw new ArgumentNullException(nameof(input))
-			: input.Length == 0
-			? Enumerable.Empty<StringSegment>()
-			: AsSegmentsCore(pattern, input);
-
-		static IEnumerable<StringSegment> AsSegmentsCore(Regex pattern, string input)
-		{
-			var match = pattern.Match(input);
-			while (match.Success)
-			{
-				yield return new(input, match.Index, match.Length);
-				match = match.NextMatch();
-			}
-		}
-	}
 
 	#region Numeric string formatting.
 	/// <summary>
