+++ Source/StringBuilderHelper.cs
@@ -8,10 +8,10 @@ namespace Open.Text;
 [SuppressMessage("Usage", "CA2225:Operator overloads have named alternates", Justification = "<Pending>")]
 public readonly record struct StringBuilderHelper
 {
-    /// <summary>
-    /// The underlying <see cref="Builder"/>.
-    /// </summary>
-    public StringBuilder Builder { get; }
+	/// <summary>
+	/// The underlying <see cref="Builder"/>.
+	/// </summary>
+	public StringBuilder Builder { get; }
 
 	/// <inheritdoc cref="StringBuilder.ToString()"/>
 	public override string ToString()
@@ -33,16 +33,16 @@ public readonly record struct StringBuilderHelper
 	/// Constructs a new <see cref="StringBuilderHelper"/> with a <see cref="StringBuilder"/> of the <paramref name="initialCapacity"/>.
 	/// </summary>
 	public StringBuilderHelper(int initialCapacity)
-        : this(new StringBuilder(initialCapacity)) { }
+		: this(new StringBuilder(initialCapacity)) { }
 
-    /// <summary>
-    /// Appends the characters to the underlying <see cref="StringBuilder"/>.
-    /// </summary>
-    public static StringBuilderHelper Add(StringBuilderHelper helper, string characters)
-    {
+	/// <summary>
+	/// Appends the characters to the underlying <see cref="StringBuilder"/>.
+	/// </summary>
+	public static StringBuilderHelper Add(StringBuilderHelper helper, string characters)
+	{
 		helper.Builder.Append(characters);
-        return helper;
-    }
+		return helper;
+	}
 
 	/// <summary>
 	/// Appends the <paramref name="character"/> to the underlying <see cref="StringBuilder"/>.
@@ -52,114 +52,114 @@ public readonly record struct StringBuilderHelper
 
 	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
 	public static StringBuilderHelper operator +(StringBuilderHelper helper, string characters)
-        => Add(helper, characters);
+		=> Add(helper, characters);
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, StringSegment characters)
-    {
-        helper.Builder.AppendSegment(characters);
-        return helper;
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, StringSegment characters)
+	{
+		helper.Builder.AppendSegment(characters);
+		return helper;
 	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, ReadOnlySpan<char> characters)
-    {
-        helper.Builder.Append(characters);
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, ReadOnlySpan<char> characters)
+	{
+		helper.Builder.Append(characters);
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, Span<char> characters)
-    {
-        helper.Builder.Append(characters);
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, Span<char> characters)
+	{
+		helper.Builder.Append(characters);
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, char[] characters)
-    {
-        if (characters is null) return helper;
-        helper.Builder.Append(characters.AsSpan());
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, char[] characters)
+	{
+		if (characters is null) return helper;
+		helper.Builder.Append(characters.AsSpan());
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, ReadOnlyMemory<char> characters)
-    {
-        helper.Builder.Append(characters.Span);
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, ReadOnlyMemory<char> characters)
+	{
+		helper.Builder.Append(characters.Span);
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, Memory<char> characters)
-    {
-        helper.Builder.Append(characters.Span);
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, Memory<char> characters)
+	{
+		helper.Builder.Append(characters.Span);
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, ArraySegment<char> characters)
-    {
-        helper.Builder.Append(characters.AsSpan());
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, ArraySegment<char> characters)
+	{
+		helper.Builder.Append(characters.AsSpan());
+		return helper;
+	}
 
-    /// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
-    public static StringBuilderHelper operator +(StringBuilderHelper helper, IEnumerable<char> characters)
-    {
-        if (characters is null) return helper;
-        var sb = helper.Builder;
-        foreach(var c in characters)
-            sb.Append(c);
-        return helper;
-    }
+	/// <inheritdoc cref="Add(StringBuilderHelper, string)"/>
+	public static StringBuilderHelper operator +(StringBuilderHelper helper, IEnumerable<char> characters)
+	{
+		if (characters is null) return helper;
+		var sb = helper.Builder;
+		foreach (var c in characters)
+			sb.Append(c);
+		return helper;
+	}
 
-    /// <summary>
-    /// Creates a new <see cref="StringBuilderHelper"/> instance beginning with the specified <paramref name="sequence"/>.
-    /// </summary>
-    internal static StringBuilderHelper NewFrom<T>(T sequence)
-        => throw new NotImplementedException();
+	/// <summary>
+	/// Creates a new <see cref="StringBuilderHelper"/> instance beginning with the specified <paramref name="sequence"/>.
+	/// </summary>
+	internal static StringBuilderHelper NewFrom<T>(T sequence)
+		=> throw new NotImplementedException();
 
 	/// <inheritdoc cref="NewFrom{T}(T)"/>
 	[ExcludeFromCodeCoverage]
 	public static implicit operator StringBuilderHelper(string sequence)
-        => string.IsNullOrEmpty(sequence) ? new() : new(new StringBuilder(sequence));
+		=> string.IsNullOrEmpty(sequence) ? new() : new(new StringBuilder(sequence));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(char value)
-        => new(new StringBuilder(1).Append(value));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(char value)
+		=> new(new StringBuilder(1).Append(value));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(StringSegment value)
-        => new(new StringBuilder(value.Length).AppendSegment(value));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(StringSegment value)
+		=> new(new StringBuilder(value.Length).AppendSegment(value));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(ReadOnlySpan<char> value)
-        => new(new StringBuilder(value.Length).Append(value));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(ReadOnlySpan<char> value)
+		=> new(new StringBuilder(value.Length).Append(value));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(ReadOnlyMemory<char> value)
-        => new(new StringBuilder(value.Length).Append(value.Span));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(ReadOnlyMemory<char> value)
+		=> new(new StringBuilder(value.Length).Append(value.Span));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(Memory<char> value)
-        => new(new StringBuilder(value.Length).Append(value.Span));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(Memory<char> value)
+		=> new(new StringBuilder(value.Length).Append(value.Span));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(ArraySegment<char> value)
-        => new(new StringBuilder(value.Count).Append(value.AsSpan()));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(ArraySegment<char> value)
+		=> new(new StringBuilder(value.Count).Append(value.AsSpan()));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(char[] value)
-        => value is null ? new() : new(new StringBuilder(value.Length).Append(value.AsSpan()));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(char[] value)
+		=> value is null ? new() : new(new StringBuilder(value.Length).Append(value.AsSpan()));
 
-    /// <inheritdoc cref="NewFrom{T}(T)"/>
-    public static implicit operator StringBuilderHelper(StringBuilder value)
-        => value is null ? new() : new(new StringBuilder(value.Length).Append(value));
+	/// <inheritdoc cref="NewFrom{T}(T)"/>
+	public static implicit operator StringBuilderHelper(StringBuilder value)
+		=> value is null ? new() : new(new StringBuilder(value.Length).Append(value));
 
-    /// <summary>
-    /// Converts the <see cref="StringBuilderHelper"/> instance to a string.
-    /// </summary>
-    public static implicit operator string(StringBuilderHelper value)
-        => value.Builder.ToString();
+	/// <summary>
+	/// Converts the <see cref="StringBuilderHelper"/> instance to a string.
+	/// </summary>
+	public static implicit operator string(StringBuilderHelper value)
+		=> value.Builder.ToString();
 }
