+++ Source/_Global.cs
@@ -1,8 +1,8 @@
 ﻿global using Microsoft.Extensions.Primitives;
-global using System.Runtime.CompilerServices;
 global using System.Diagnostics;
 global using System.Diagnostics.CodeAnalysis;
 global using System.Diagnostics.Contracts;
 global using System.Globalization;
 global using System.Reflection;
+global using System.Runtime.CompilerServices;
 global using System.Text.RegularExpressions;
