+++ Tests/SplitTests.cs
@@ -1,5 +1,5 @@
-﻿using System.Text.RegularExpressions;
-using System.ComponentModel;
+﻿using System.ComponentModel;
+using System.Text.RegularExpressions;
 
 namespace Open.Text.Tests;
