+++ Tests/_Global.cs
@@ -1,3 +1,3 @@
 ﻿global using FluentAssertions;
-global using Xunit;
 global using System.Diagnostics.CodeAnalysis;
+global using Xunit;
